refactor: GameManager class 분리
